# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: MoLock API
  version: 2.0.0
  description: '6인 랜덤 매칭 라이브 캠스터디 서비스 MoLock의 백엔드 API입니다. 구 모락(사진 인증 챌린지)에서 라이브 세션으로 피벗했습니다.


    시간만 고르면 같은 목표 시간의 6명이 자동으로 묶여 라이브 세션이 열리고,

    세션이 끝날 때까지 자리를 지키면 완주로 인정돼 포인트와 Streak가 쌓입니다.

    캠 영상은 서버로 오지 않고 각자 기기에서만 분석되며, 서버는 자리비움 여부와 시각만 받습니다.


    ## 인증


    로그인으로 받은 토큰을 `Authorization: Bearer {accessToken}` 헤더에 넣어 보냅니다.

    토큰은 24시간 유효하며 갱신(refresh) 토큰은 없습니다. 만료되면 다시 로그인해야 합니다.


    ## 오류 응답


    모든 실패 응답은 아래 형태로 통일되어 있습니다.


    ```json

    {"error": {"code": "SESSION_NOT_FOUND", "message": "존재하지 않는 세션입니다.", "details": null}}

    ```


    화면 분기는 `message`가 아니라 **`code`** 로 하세요. 문구는 다듬어질 수 있지만 코드는 바뀌지 않습니다.

    에러 코드 하나에는 HTTP status 하나만 대응합니다. 같은 코드가 상황에 따라 다른 status로 오지 않습니다.


    ## 어느 API에서나 날 수 있는 오류


    아래 두 가지는 모든 경로에 해당하므로 개별 API 설명에서는 생략했습니다.


    | 상태 | 코드 | 언제 | 화면 처리 |

    | --- | --- | --- | --- |

    | 404 | `ENDPOINT_NOT_FOUND` | 없는 주소를 호출했을 때 | 개발 중 오류. 경로를 확인하세요 |

    | 500 | `INTERNAL_SERVER_ERROR` | 서버 오류 | 잠시 후 다시 시도하도록 안내 |


    인증·회원상태·제재·연령 게이트의 401·403은 각 API의 응답 설명에 적혀 있습니다.


    ## 목록 응답


    목록은 `{content[], page, size, totalElements, totalPages}` 형태입니다.

    `page`는 0부터, `size`는 기본 20·최대 50이며 초과하면 400 `VALIDATION_FAILED`입니다.


    ## 시각 표기


    모든 시각은 ISO-8601(`+09:00`), 일자는 `YYYY-MM-DD`입니다.

    '
servers:
- url: http://localhost:8080/api
  description: 개발
- url: https://api.molock.example.com/api
  description: 운영 (배포 시 확정)
...
security:
- bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 로그인 응답의 `accessToken`을 넣습니다. 24시간 유효합니다.
  schemas:
    ErrorResponse:
      type: object
      description: 모든 실패 응답의 공통 형태
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              description: 화면 분기에 쓰는 값
              example: SESSION_NOT_FOUND
            message:
              type: string
              description: 사용자에게 그대로 보여줘도 되는 문구
              example: 존재하지 않는 세션입니다.
            details:
              type: object
              nullable: true
              additionalProperties: true
              description: 추가 정보. 없으면 `null`
              example: null
    PageMeta:
      type: object
      description: 목록 응답의 공통 페이지 정보
      required:
      - page
      - size
      - totalElements
      - totalPages
      properties:
        page:
          type: integer
          description: 현재 페이지. 0부터 시작
          example: 0
        size:
          type: integer
          description: 한 페이지 항목 수
          example: 20
        totalElements:
          type: integer
          description: 전체 항목 수
          example: 37
        totalPages:
          type: integer
          description: 전체 페이지 수
          example: 2
...
    SessionStatus:
      type: string
      description: '세션 상태. LIVE 진행 중 · ENDED 종료 · CANCELLED(v1에 전이 경로 없음)'
      enum:
      - LIVE
      - ENDED
      - CANCELLED
      example: LIVE
...
    ParticipantStatus:
      type: string
      description: '참가자 상태. ACTIVE 참여 중 · PAUSED 화장실 모드 · LEFT 퇴장 · EVICTED 경고 3회 퇴출'
      enum:
      - ACTIVE
      - PAUSED
      - LEFT
      - EVICTED
      example: ACTIVE
...
    AppealStatus:
      type: string
      description: '이의 신청 상태. PENDING 접수 · ACCEPTED 인용 · REJECTED 기각'
      enum:
      - PENDING
      - ACCEPTED
      - REJECTED
      example: PENDING
    AppealDecision:
      type: string
      description: 이의 처리 결정값. PENDING으로는 되돌릴 수 없습니다
      enum:
      - ACCEPTED
      - REJECTED
      example: ACCEPTED
...
    DecidedBy:
      type: string
      description: '판정 주체. AI 자동 판정 · ADMIN 운영자'
      enum:
      - AI
      - ADMIN
      example: ADMIN
...
    TargetMinutes:
      type: integer
      description: 하루 목표 시간(분). 대기열이 이 값으로 갈립니다
      enum:
      - 60
      - 120
      - 180
      - 240
      example: 120
...
    AppealQueueItem:
      type: object
      required:
      - appealId
      - evictionId
      - memberId
      - nickname
      - sessionId
      - warningCount
      - status
      - overdue
      - createdAt
      - slaDueAt
      properties:
        appealId:
          type: integer
          format: int64
        evictionId:
          type: integer
          format: int64
        memberId:
          type: integer
          format: int64
        nickname:
          type: string
        sessionId:
          type: integer
          format: int64
        warningCount:
          type: integer
        status:
          $ref: '#/components/schemas/AppealStatus'
        overdue:
          type: boolean
          description: 신고 케이스와 같은 규칙의 파생값입니다 (`status=PENDING AND slaDueAt < now`)
        createdAt:
          type: string
          format: date-time
        slaDueAt:
          type: string
          format: date-time
    AppealPageResponse:
      allOf:
      - $ref: '#/components/schemas/PageMeta'
      - type: object
        required:
        - content
        properties:
          content:
            type: array
            items:
              $ref: '#/components/schemas/AppealQueueItem'
      example:
        content:
        - appealId: 41
          evictionId: 77
          memberId: 1042
          nickname: 익명 치타37
          sessionId: 5501
          warningCount: 3
          status: PENDING
          overdue: false
          createdAt: '2026-08-12T11:20:00+09:00'
          slaDueAt: '2026-08-15T11:20:00+09:00'
        page: 0
        size: 20
        totalElements: 5
        totalPages: 1
    AppealProcessRequest:
      type: object
      required:
      - decision
      properties:
        decision:
          $ref: '#/components/schemas/AppealDecision'
        note:
          type: string
          maxLength: 1000
          nullable: true
          description: 관리자 처리 사유. 신청자의 `reasonText`를 덮어쓰지 않습니다
      example:
        decision: ACCEPTED
        note: 자리비움 이벤트 로그상 카메라 각도 문제로 확인. 퇴출 취소.
    AppealProcessResponse:
      type: object
      required:
      - appealId
      - status
      - decidedBy
      - decidedAt
      properties:
        appealId:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/AppealStatus'
        decidedBy:
          $ref: '#/components/schemas/DecidedBy'
        decidedAt:
          type: string
          format: date-time
        pointRefunded:
          type: integer
          nullable: true
          description: 인용 시 역분개로 되돌린 포인트. 기각이면 `0`입니다. 되돌릴 차감이 아직 원장에 없으면 인용이어도 `0`입니다
        sessionCompletedRestored:
          type: boolean
          nullable: true
          description: 완주 소급 재판정 결과. 세션이 아직 진행 중이면 소급하지 않으므로 `false`입니다
        streakAfter:
          type: integer
          nullable: true
          description: 재계산된 Streak
      example:
        appealId: 41
        status: ACCEPTED
        decidedBy: ADMIN
        decidedAt: '2026-08-13T09:10:00+09:00'
        pointRefunded: 300
        sessionCompletedRestored: true
        streakAfter: 5
    AdminSessionParticipant:
      type: object
      required:
      - memberId
      - nickname
      - status
      - warningCount
      - paused
      properties:
        memberId:
          type: integer
          format: int64
        nickname:
          type: string
        status:
          $ref: '#/components/schemas/ParticipantStatus'
        warningCount:
          type: integer
        paused:
          type: boolean
    AdminSessionSummary:
      type: object
      required:
      - sessionId
      - status
      - targetMinutes
      - startedAt
      - endsAt
      - activeCount
      - pausedCount
      - leftCount
      - evictedCount
      - participants
      properties:
        sessionId:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/SessionStatus'
        targetMinutes:
          $ref: '#/components/schemas/TargetMinutes'
        startedAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        activeCount:
          type: integer
        pausedCount:
          type: integer
        leftCount:
          type: integer
        evictedCount:
          type: integer
        participants:
          type: array
          items:
            $ref: '#/components/schemas/AdminSessionParticipant'
    AdminSessionPageResponse:
      allOf:
      - $ref: '#/components/schemas/PageMeta'
      - type: object
        required:
        - content
        properties:
          content:
            type: array
            items:
              $ref: '#/components/schemas/AdminSessionSummary'
      example:
        content:
        - sessionId: 5501
          status: LIVE
          targetMinutes: 120
          startedAt: '2026-08-12T09:00:00+09:00'
          endsAt: '2026-08-12T11:00:00+09:00'
          activeCount: 4
          pausedCount: 1
          leftCount: 0
          evictedCount: 1
          participants:
          - memberId: 1042
            nickname: 익명 치타37
            status: ACTIVE
            warningCount: 1
            paused: false
        page: 0
        size: 20
        totalElements: 3
        totalPages: 1
...
  responses:
    Unauthorized:
      description: '인증 실패

        · `UNAUTHORIZED` 토큰이 없거나 잘못됨, 또는 이미 파기된 계정 → 로그인 화면으로

        · `TOKEN_EXPIRED` 토큰이 24시간을 넘김 → 로그인 화면으로

        '
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    AdminForbidden:
      description: "관리 권한 없음\n· `FORBIDDEN_ROLE` 관리자가 아닌 계정이 호출함 → **관리 탭 자체를 숨깁니다.**\n\
        \  일반 사용자 앱에서는 이 응답을 받을 일이 없어야 합니다.\n· `WITHDRAWAL_PENDING` 탈퇴 신청 중인 계정\n"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN_ROLE
              message: 관리자 권한이 필요합니다.
              details: null
    ValidationFailed:
      description: '`VALIDATION_FAILED` 요청 값이 형식·허용 범위에 맞지 않음 → 입력 오류 표시'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PageValidationFailed:
      description: '`VALIDATION_FAILED` 페이지 파라미터가 범위를 벗어남(`size` 최대 50)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  parameters:
...
    AppealId:
      name: appealId
      in: path
      required: true
      description: 이의 신청 ID
      schema:
        type: integer
        format: int64
      example: 41
...
    Page:
      name: page
      in: query
      required: false
      description: 조회할 페이지. 0부터 시작
      schema:
        type: integer
        default: 0
        minimum: 0
    Size:
      name: size
      in: query
      required: false
      description: 한 페이지 항목 수. 최대 50
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 50
paths:
...
  /admin/appeals:
    get:
      tags:
      - 관리자
      summary: AD-5 이의 큐
      operationId: getAdminAppeals
      description: "퇴출 이의 신청 처리 큐입니다.\n\n- `overdue`는 신고 케이스와 같은 규칙의 파생값입니다 — 처리 대기\
        \ 상태이면서 SLA 기한이 지난 건.\n"
      parameters:
      - name: status
        in: query
        required: false
        description: 이의 상태 필터
        schema:
          $ref: '#/components/schemas/AppealStatus'
      - name: overdue
        in: query
        required: false
        description: SLA 초과 건만 보기
        schema:
          type: boolean
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppealPageResponse'
        '400':
          $ref: '#/components/responses/PageValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/AdminForbidden'
  /admin/appeals/{appealId}:
    patch:
      tags:
      - 관리자
      summary: AD-6 이의 처리
      operationId: processAppeal
      description: "이의를 인용하거나 기각합니다.\n\n- **인용해도 퇴출 기록 자체는 남깁니다(감사 추적).** 참가자 상태는\
        \ 퇴출로 유지하고 취소 사실은 별도 시각으로 표현합니다.\n- 인용 시 패널티는 역분개로 되돌아가고, 원래의 차감 행은 지우지 않습니다.\n\
        - 인용 시 완주를 소급 재판정합니다 — 퇴출이 없었다면 세션 종료까지 참가한 것으로 보아 완주 처리하고 포인트를 지급합니다.\n \
        \ Streak도 함께 재계산되며 목표 달성 조건을 채우면 달성 처리됩니다.\n  단, 세션이 아직 진행 중이면 완주 여부가 정해지지\
        \ 않았으므로 소급하지 않습니다(`sessionCompletedRestored=false`).\n- 되돌릴 차감이 아직 원장에 없으면 역분개도\
        \ 만들지 않습니다(`pointRefunded=0`). 역분개는 기존 기록을 뒤집는 것이지 새 지급이 아닙니다.\n- 재매칭 쿨다운은\
        \ 인용 시점부터 해제됩니다.\n- 기각이면 상태와 메모만 기록하고 원복은 일어나지 않습니다.\n"
      parameters:
      - $ref: '#/components/parameters/AppealId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppealProcessRequest'
      responses:
        '200':
          description: 처리 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppealProcessResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/AdminForbidden'
        '404':
          description: '`APPEAL_NOT_FOUND` 존재하지 않는 이의'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: '`ALREADY_PROCESSED` 이미 처리된 이의'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/sessions:
    get:
      tags:
      - 관리자
      summary: AD-7 진행 중 세션 모니터
      operationId: getAdminSessions
      description: "진행 중인 세션과 참가자 상태를 봅니다.\n\n- **관리자에게도 영상은 제공하지 않습니다.** 저장하지 않으므로\
        \ 존재하지 않습니다.\n- `status`를 생략하면 종료된 세션까지 함께 나옵니다.\n- 이의가 인용된 퇴출도 `evictedCount`에\
        \ 남습니다 — 참가자 상태는 퇴출로 유지되고 취소 사실은 별도 시각으로 표현하기 때문입니다.\n- 조회 전용입니다. 관리자가 세션에\
        \ 개입하는 경로는 v1 범위 밖입니다.\n"
      parameters:
      - name: status
        in: query
        required: false
        description: 세션 상태 필터
        schema:
          $ref: '#/components/schemas/SessionStatus'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSessionPageResponse'
        '400':
          $ref: '#/components/responses/PageValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/AdminForbidden'
...